backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.2
nltk==3.8.1
yfinance==0.2.31
beautifulsoup4==4.12.2
requests==2.31.0
torch==2.1.0
transformers==4.34.1
python-dotenv==1.0.0
jwt==1.3.1
starlette==0.27.0